requirements.txt
gunicorn==23.0.0
Flask==3.0.3
requests==2.32.3
Pillow==10.4.0
numpy==2.0.1
opencv-python-headless==4.11.0.86
aiohttp==3.9.1
